Handle case when `glue()` is called with empty inputs
glue 1.5.0 changes the behavior when called with completely empty
inputs.

Previous versions of glue returned `""`, e.g. a length 1 character
vector with an empty string as the value.

glue 1.5.0 instead returns `character(0)` vector, a 0 length character
vector.

The new behavior is more consistent with `paste0()`, which also returns
a 0 length character vector on empty input.

This change caused an example to fail with glue 1.5.0, which is fixed by
the changes in this commit.